+++ src/components/sort-component.js
@@ -11,7 +11,7 @@ const createSortTemplate = (sortType = SortType.DEFAULT) => {
   `);
 };
 
-export default class Sort extends AbstractSmartComponent {
+export default class SortComponent extends AbstractSmartComponent {
   constructor() {
     super();
     this._currentSortType = SortType.DEFAULT;
