+++ src/controllers/page.js
@@ -2,14 +2,14 @@ import API from "../api";
 import {FilmsQuantity, NoDataStatus, Pages, SortType} from "../const";
 import {remove, render} from "../utils/render";
 import {getProfileRank, getWatchedFilmsByPeriod, sortFilms} from "../utils/common";
-import Profile from "../components/profile";
-import Sort from "../components/sort";
-import Statistic from "../components/statistic";
-import Films from "../components/films";
+import ProfileComponent from "../components/profile-component";
+import SortComponent from "../components/sort-component";
+import StatisticComponent from "../components/statistic-component";
+import FilmsComponent from "../components/films-component";
 import FilmsController from "./films";
 import FilmsModel from "../models/films";
 import FilterController from "../controllers/filter";
-import FooterStatistics from "../components/footer-statistics";
+import FooterStatisticsComponent from "../components/footer-statistics-component";
 import FilmsNoData from "../components/films-no-data";
 
 const AUTHORIZATION = `Basic gfjdoHFJDL59fdsfds7`;
@@ -24,12 +24,12 @@ export default class PageController {
     this._siteFooterStatisticsElement = document.querySelector(`.footer__statistics`);
     this._showingFilmsCount = FilmsQuantity.SHOWING_ON_START;
     this._filmsModel = new FilmsModel();
-    this._profileComponent = new Profile();
-    this._sortComponent = new Sort();
-    this._filmsComponent = new Films();
+    this._profileComponent = new ProfileComponent();
+    this._sortComponent = new SortComponent();
+    this._filmsComponent = new FilmsComponent();
     this._filmsController = null;
-    this._statistic = new Statistic();
-    this._footerComponent = new FooterStatistics(0);
+    this._statistic = new StatisticComponent();
+    this._footerComponent = new FooterStatisticsComponent(0);
     this._onPageChange = this._onPageChange.bind(this);
     this._onSortTypeChange = this._onSortTypeChange.bind(this);
     this._onStatsChange = this._onStatsChange.bind(this);
